fix(lint): extract CopyButton component and fix TypeScript ref types
