+++ src/components/RawEventJson.tsx
@@ -3,9 +3,7 @@
 import { NDKEvent } from '@nostr-dev-kit/ndk';
 import { Highlight, themes, type RenderProps } from 'prism-react-renderer';
 import { toPlainEvent } from '@/lib/toPlainEvent';
-import { useEffect, useRef, useState } from 'react';
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import { faCopy, faCheck } from '@fortawesome/free-solid-svg-icons';
+import CopyButton from '@/components/CopyButton';
 
 type Props = {
   event: NDKEvent | null | undefined;
@@ -41,29 +39,7 @@ export default function RawEventJson({ event, loading = false, className, title,
         return headerTitle ? (
         <div className="text-xs text-gray-300 mb-2 flex items-center justify-between">
           <span className="font-semibold">{headerTitle}</span>
-          {(() => {
-            const [copied, setCopied] = useState(false);
-            const timerRef = useRef<ReturnType<typeof setTimeout> | null>(null);
-            useEffect(() => () => { if (timerRef.current) clearTimeout(timerRef.current); }, []);
-            const handleCopy = async (e: React.MouseEvent<HTMLButtonElement>) => {
-              e.preventDefault();
-              try { await navigator.clipboard.writeText(json); } catch {}
-              setCopied(true);
-              if (timerRef.current) clearTimeout(timerRef.current);
-              timerRef.current = setTimeout(() => setCopied(false), 1500);
-            };
-            return (
-              <button
-                type="button"
-                title={copied ? 'Copied' : 'Copy JSON'}
-                aria-label={copied ? 'Copied' : 'Copy JSON'}
-                className="w-6 h-6 rounded-md border border-[#3d3d3d] text-gray-300 hover:bg-[#2a2a2a] flex items-center justify-center"
-                onClick={handleCopy}
-              >
-                <FontAwesomeIcon icon={copied ? faCheck : faCopy} className="text-xs" />
-              </button>
-            );
-          })()}
+          <CopyButton text={json} title="Copy JSON" />
         </div>
         ) : null;
       })()}
